// types/marketplace-api-schema.ts
export interface CreateAppModel {
  /**
   * Sets the full name of the app
   */
  name?: string
  /**
   * Sets the description of the app
   */
  description?: string
  /**
   * Sets a short summary of the app
   */
  summary?: string
  /**
   * Sets the email address of the developer's helpdesk responsible for providing support of the app
   */
  supportEmail?: string
  /**
   * Sets the telephone number of the developer's helpdesk responsible for providing support of the app
   */
  telephone?: string
  /**
   * Sets the home page of the developer, or the application/product specific page on the developer's website
   */
  homePage?: string
  /**
   * Sets the apps launch uri
   */
  launchUri?: string
  /**
   * Sets the unique identifer of the developer registering the app
   */
  developerId?: string // uuid
  /**
   * Sets the unique identifer of the apps category
   */
  categoryId?: string
  /**
   * Sets a flag determining whether or not the app is featured
   */
  isFeatured?: boolean
  /**
   * Sets the flag determining whether or not the app will appear in the marketplace
   */
  isDirectApi?: boolean
  /**
   * Sets the application authorisation flow type (authorisationCode/clientCredentials)
   */
  authFlow?: string
  /**
   * Sets the list of scope keys required for this app to run
   */
  scopes?: string[]
  /**
   * Sets the location url of the app icon image
   */
  iconImageUrl?: string
  /**
   * Sets the location url of the first app screenshot image
   */
  screen1ImageUrl?: string
  /**
   * Sets the location url of the second (optional) app screenshot image
   */
  screen2ImageUrl?: string
  /**
   * Sets the location url of the third (optional) app screenshot image
   */
  screen3ImageUrl?: string
  /**
   * Sets the location url of the fourth (optional) app screenshot image
   */
  screen4ImageUrl?: string
  /**
   * Sets the location url of the fifth (optional) app screenshot image
   */
  screen5ImageUrl?: string
}
// ...
export interface ScopeModel {
  /**
   * Gets the name of the scope
   */
  name?: string
  /**
   * Gets the description of the scope
   */
  description?: string
}